requirements.txt
aiofile==3.3.3
aiohttp==3.7.3
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
CacheControl==0.12.6
caio==0.6.3
certifi==2019.11.28
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.3.1
daphne==3.0.1
distlib==0.3.0
distro==1.4.0
Django==3.1.5
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==1.0.2
multidict==5.1.0
numpy==1.19.5
packaging==20.3
pep517==0.8.2
priority==1.3.0
progress==1.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
websockets==8.1
yarl==1.6.3
zope.interface==5.2.0
nest_asyncio==1.5.1